refactor: extract category screen components
